worked on promote ads section, we now have intact state, working on the admin helped speed things up and also thanks to ChatGPT
