+++ lib/screens/Home/Screens/Account/ManageProduct/Specifications/_spec_item_model.dart
@@ -1,0 +1,8 @@
+class SpecItemModel {
+  final String id;
+  final String? specItemTitle;
+  final String? specItemValue;
+
+  const SpecItemModel(
+      {required this.id, this.specItemTitle, this.specItemValue});
+}
