+++ lib/shared_components/utilities/formart_to_currency_naira.dart
@@ -1,0 +1,10 @@
+import 'package:intl/intl.dart';
+
+String formatToNairaCurrency(int number) {
+  final formatter = NumberFormat.currency(
+    locale: 'en_NG',
+    symbol: '₦',
+    decimalDigits: 0,
+  );
+  return formatter.format(number);
+}
